+++ components/MapScreenComponents/CustomSpotModal.js
@@ -6,18 +6,22 @@ import {
   Modal,
   TextInput,
   TouchableOpacity,
+  ScrollView,
 } from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
 
-const CustomSpotModal = ({visible, onClose, onSubmit, spot, onSpotChange}) => {
-  const emojis = ['📍', '🎯', '⭐', '🎪', '🎭', '🎡'];
+const CustomSpotModal = ({
+  visible,
+  newSpot,
+  onClose,
+  onCreateSpot,
+  onSpotChange,
+}) => {
+  const emojis = [
+    '📍', '🎯', '⭐', '🎪', '🎭', '🎡', '🎢', '🎨',
+    '🎰', '🍽️', '🏛️', '🏰', '🌟', '💫', '🌺', '🌴',
+  ];
 
-  const handleTextChange = (field, value) => {
-    onSpotChange({
-      ...spot,
-      [field]: value,
-    });
-  };
   return (
     <Modal
       animationType="slide"
@@ -31,25 +35,28 @@ const CustomSpotModal = ({visible, onClose, onSubmit, spot, onSpotChange}) => {
           <Text style={styles.modalTitle}>Create New Spot</Text>
 
           <View style={styles.emojiSelector}>
-            {emojis.map(emoji => (
-              <TouchableOpacity
-                key={emoji}
-                onPress={() => handleTextChange('emoji', emoji)}
-                style={[
-                  styles.emojiOption,
-                  spot?.emoji === emoji && styles.selectedEmoji,
-                ]}>
-                <Text style={styles.emojiText}>{emoji}</Text>
-              </TouchableOpacity>
-            ))}
+            <ScrollView horizontal showsHorizontalScrollIndicator={false}>
+              {emojis.map(emoji => (
+                <TouchableOpacity
+                  key={emoji}
+                  onPress={() => onSpotChange({...newSpot, emoji})}
+                  style={[
+                    styles.emojiOption,
+                    newSpot.emoji === emoji && styles.selectedEmoji,
+                  ]}>
+                  <Text style={styles.emojiText}>{emoji}</Text>
+                </TouchableOpacity>
+              ))}
+              <View style={{width: 20}} />
+            </ScrollView>
           </View>
 
           <TextInput
             style={styles.input}
             placeholder="Spot Name"
             placeholderTextColor="rgba(255,255,255,0.5)"
-            value={spot?.name}
-            onChangeText={text => handleTextChange('name', text)}
+            value={newSpot.name}
+            onChangeText={text => onSpotChange({...newSpot, name: text})}
           />
 
           <TextInput
@@ -58,16 +65,20 @@ const CustomSpotModal = ({visible, onClose, onSubmit, spot, onSpotChange}) => {
             placeholderTextColor="rgba(255,255,255,0.5)"
             multiline
             numberOfLines={4}
-            value={spot?.description}
-            onChangeText={text => handleTextChange('description', text)}
+            value={newSpot.description}
+            onChangeText={text => onSpotChange({...newSpot, description: text})}
           />
 
           <View style={styles.modalButtons}>
-            <TouchableOpacity style={styles.cancelButton} onPress={onClose}>
+            <TouchableOpacity
+              style={styles.cancelButton}
+              onPress={onClose}>
               <Text style={styles.buttonText}>Cancel</Text>
             </TouchableOpacity>
 
-            <TouchableOpacity style={styles.createButton} onPress={onSubmit}>
+            <TouchableOpacity
+              style={styles.createButton}
+              onPress={onCreateSpot}>
               <Text style={styles.buttonText}>Create</Text>
             </TouchableOpacity>
           </View>
@@ -83,59 +94,77 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     backgroundColor: 'rgba(0,0,0,0.5)',
-    padding: 20,
+    // paddingHorizontal: 10,
   },
   modalContent: {
     width: '100%',
-    padding: 20,
-    borderRadius: 15,
+    maxWidth: '90%', // Added maxWidth to control modal size
+    // padding: 20,
+    borderRadius: 20,
+    backgroundColor: '#2B3467',
     alignItems: 'center',
   },
   modalTitle: {
-    fontSize: 24,
+    fontSize: 28,
     fontWeight: 'bold',
     color: '#fff',
-    marginBottom: 20,
+    textAlign: 'center',
+    marginBottom: 24,
     textShadowColor: '#00ff00',
     textShadowOffset: {width: 1, height: 1},
-    textShadowRadius: 3,
+    textShadowRadius: 10,
+    marginVertical: 10,
+    paddingHorizontal: 30,
   },
   input: {
-    width: '100%',
+    // width: '100%', // Ensure input takes full width of parent
     backgroundColor: 'rgba(255,255,255,0.1)',
-    borderRadius: 10,
+    borderRadius: 15,
     padding: 15,
-    marginBottom: 15,
+    marginBottom: 16,
     color: '#fff',
     borderWidth: 1,
-    borderColor: 'rgba(0,255,0,0.3)',
+    borderColor: '#00ff00',
+    fontSize: 16,
+    marginRight: 10,
+    marginLeft: 10,
+    width: '90%',
   },
   textArea: {
     height: 100,
     textAlignVertical: 'top',
   },
   emojiSelector: {
+    // width: '100%', // Added width
     flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'center',
-    marginBottom: 20,
+    justifyContent: 'space-between',
+    marginBottom: 24,
+    marginHorizontal: 20,
   },
   emojiOption: {
-    padding: 10,
-    borderRadius: 25,
+    width: 44,
+    height: 44,
+    borderRadius: 22,
+    backgroundColor: 'rgba(255,255,255,0.1)',
+    justifyContent: 'center',
+    alignItems: 'center',
     margin: 5,
   },
   selectedEmoji: {
     backgroundColor: 'rgba(0,255,0,0.2)',
+    borderWidth: 2,
+    borderColor: '#00ff00',
   },
   emojiText: {
     fontSize: 24,
   },
   modalButtons: {
+    // width: '100%', // Added width
     flexDirection: 'row',
     justifyContent: 'space-between',
-    width: '100%',
-    marginTop: 20,
+    marginTop: 24,
+    marginBlock: 20,
+    marginHorizontal: 10,
   },
   createButton: {
     flex: 1,
@@ -146,10 +175,10 @@ const styles = StyleSheet.create({
   },
   cancelButton: {
     flex: 1,
-    backgroundColor: 'rgba(255,255,255,0.2)',
+    backgroundColor: 'rgba(255,255,255,0.15)',
     padding: 15,
-    borderRadius: 10,
-    marginRight: 10,
+    borderRadius: 15,
+    marginRight: 8,
   },
   buttonText: {
     color: '#fff',
