+++ components/MapScreenComponents/index.js
@@ -1,3 +1,12 @@
 import SpotNotice from './SpotNotice';
-
-export {SpotNotice};
+import RoutingControls from './RoutingControls';
+import CustomMarkers from './CustomMarkers';
+import CustomSpotModal from './CustomSpotModal';
+import LasVegasButton from './LasVegasButton';
+export {
+  SpotNotice,
+  RoutingControls,
+  CustomMarkers,
+  CustomSpotModal,
+  LasVegasButton,
+};
